Moved generator producer methods from org.scalatest.prop package object to CommonGenerators, a trait from which GeneratorDrivenPropertyChecks inherits.
