+++ scalatest-test/src/test/scala/org/scalatest/prop/GeneratorSpec.scala
@@ -88,17 +88,17 @@ class GeneratorSpec extends FunSpec with Matchers {
       tupCanonicals shouldBe expectedTupCanonicals
     }
     it("should offer a filter method so that pattern matching can be used in for expressions with Generator generators") {
-      """for ((a, b) <- tuple2s[String, Int]) yield (b, a)""" should compile
+      """for ((a, b) <- CommonGenerators.tuple2s[String, Int]) yield (b, a)""" should compile
       case class Person(name: String, age: Int)
-      val persons = gen(Person) { p => (p.name, p.age) }
+      val persons = CommonGenerators.gen(Person) { p => (p.name, p.age) }
       """for (Person(a, b) <- persons) yield (b, a)""" should compile
     }
     it("should offer a filter method that throws an exception if too many objects are filtered out") {
-      val doNotDoThisAtHome = ints.filter(i => i == 0) // Only keep zero
+      val doNotDoThisAtHome = CommonGenerators.ints.filter(i => i == 0) // Only keep zero
       a [IllegalStateException] should be thrownBy {
         doNotDoThisAtHome.next(100, Nil, Randomizer.default())
       }
-      val okToDoThisAtHome = ints.filter(i => i != 0) // Only keep non-zeros
+      val okToDoThisAtHome = CommonGenerators.ints.filter(i => i != 0) // Only keep non-zeros
       noException should be thrownBy {
         okToDoThisAtHome.next(100, Nil, Randomizer.default())
       }
